webapp/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
python-dotenv==1.0.0
beautifulsoup4==4.12.3
google-generativeai==0.3.2
pydub==0.25.1
PyAudio==0.2.14 